requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-core==0.3.17
openai==1.55.3
python-dotenv==1.0.1
streamlit==1.41.1
pandas==2.2.2
numpy==1.26.4